# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Ecosystem Core Backend API",
    "summary": "Unified RESTful API Suite for AI Data Pipelines, Storage, Database, and Worker Tasks",
    "description": "\n# AI Ecosystem Backend Service 🚀\n\nWelcome to the **AI Ecosystem Backend API** documentation. This service serves as the core orchestration backend for the AI platform.\n\n---\n\n### 🧩 Connected Components:\n* **🐘 PostgreSQL Database**: Structured data storage and ORM model persistence.\n* **📦 MinIO Object Storage**: S3-compatible file storage with object versioning.\n* **⚡ Redis & ARQ Worker**: High-performance asynchronous background task queue.\n* **🏷️ Label Studio**: Human-in-the-loop dataset annotation and task management.\n* **📊 Centralized Logger & Diagnostics**: Live system health inspection.\n\n---\n\n### 📚 Documentation & Schema Endpoints:\n* **Interactive Swagger UI**: [`/docs`](/docs)\n* **ReDoc Interactive Documentation**: [`/redoc`](/redoc)\n* **Raw OpenAPI Specification (JSON)**: [`/openapi.json`](/openapi.json)\n    ",
    "termsOfService": "https://example.com/terms/",
    "contact": {
      "name": "AI Ecosystem Development Team",
      "url": "https://github.com/Natthawut02/6710110138-AIEcosystem-Aj-Warintorn",
      "email": "natthawut@example.com"
    },
    "license": {
      "name": "MIT License",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/train/enqueue": {
      "post": {
        "tags": [
          "Trainer Worker"
        ],
        "summary": "เข้าคิวสั่งเทรนโมเดล (Token Classification)",
        "description": "รับคำสั่งเทรนโมเดล แล้ว enqueue งานเข้าคิว ARQ ชื่อ `training_queue` โดยเฉพาะ\nเพื่อให้ Trainer Worker (container แยกที่มี GPU) เป็นตัวเดียวที่หยิบงานนี้ไปทำ\n\n- **dataset_object_name**: ชื่อไฟล์ dataset ที่ต้องอัปโหลดเข้า MinIO ไว้ล่วงหน้าแล้ว\n- **model_output_name**: ชื่อโมเดลที่จะใช้บันทึกกลับไป MinIO เมื่อเทรนเสร็จ\n- **queue_time**: ถ้าระบุ งานจะไม่เริ่มจนกว่าจะถึงเวลานี้ (ใช้ ARQ `_defer_until` เบื้องหลัง)",
        "operationId": "enqueue_train_job_api_v1_train_enqueue_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainJobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "เข้าคิวสำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainJobInfo"
                }
              }
            }
          },
          "500": {
            "description": "เข้าคิวไม่สำเร็จ (Redis connection error)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/train/{job_id}": {
      "get": {
        "tags": [
          "Trainer Worker"
        ],
        "summary": "ตรวจสอบสถานะงานเทรน",
        "description": "ตรวจสอบว่างานเทรนที่ enqueue ไว้อยู่ในสถานะใด (queued / in_progress / complete)",
        "operationId": "get_train_job_status_api_v1_train__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Always false for error responses",
            "default": false
          },
          "error_code": {
            "type": "string",
            "title": "Error Code",
            "description": "Machine-readable error code",
            "default": "INTERNAL_SERVER_ERROR"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Detailed description of the error"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Additional context or validation error details"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response model."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TrainJobCreate": {
        "properties": {
          "dataset_object_name": {
            "type": "string",
            "title": "Dataset Object Name",
            "description": "ชื่อไฟล์ dataset ที่เก็บอยู่ใน MinIO (bucket: datasets) เช่น 'conll2003/train.jsonl'",
            "examples": [
              "conll2003/train.jsonl"
            ]
          },
          "model_output_name": {
            "type": "string",
            "title": "Model Output Name",
            "description": "ชื่อโมเดลที่จะใช้ตั้งตอนบันทึกกลับไป MinIO เช่น 'token-classifier-v1'",
            "examples": [
              "token-classifier-v1"
            ]
          },
          "epochs": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Epochs",
            "description": "จำนวนรอบการเทรน",
            "default": 1
          },
          "queue_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Time",
            "description": "เวลาที่ต้องการให้เริ่มเทรน (ISO 8601, เช่น 2026-09-01T22:00:00) ถ้าไม่ระบุ จะเริ่มเทรนทันทีที่ Trainer Worker ว่าง",
            "examples": [
              "2026-09-01T22:00:00"
            ]
          }
        },
        "type": "object",
        "required": [
          "dataset_object_name",
          "model_output_name"
        ],
        "title": "TrainJobCreate",
        "description": "ข้อมูลที่ Client ต้องส่งมาเพื่อขอเข้าคิวเทรนโมเดล"
      },
      "TrainJobInfo": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "queue_name": {
            "type": "string",
            "title": "Queue Name"
          },
          "scheduled_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Start"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "queue_name"
        ],
        "title": "TrainJobInfo",
        "description": "ข้อมูลตอบกลับหลัง enqueue งานเทรนสำเร็จ"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}